Move utilities to root utils directory
