+++ app/routes/admin/views/index.tsx
@@ -3,9 +3,9 @@ import { useActionData, useNavigation, useSearchParams } from "react-router";
 import { redirect } from "react-router";
 import { Toaster, toast } from "sonner";
 import { ValiError } from "valibot";
-import { assert } from "~/routes/common/utils/assert";
-import { checkSession } from "~/routes/common/utils/auth";
 import { updateContent } from "~/services/db.server";
+import { assert } from "~/utils/assert";
+import { checkSession } from "~/utils/auth";
 import { validateContentInsert } from "../../../../database/valibot-validation.js";
 import AdminDashboard from "../components/AdminDashboard";
 import { ButtonLED } from "../components/ui/button";
